instead of crafting, use place-and-dye to colorize
That is, place a piece of obsidian framed glass on the
ground, then right-click it with dye to set a color (by default,
it is clear/empty).  Dig the colored one to get back the uncolored
one and dye.

You can change the color by dyeing it again, in which case you get
back the old dye.

If the color you wield is not supported by the mod, or if the target
node is already that color, a warning message is thrown.